refactor(metagame_analysis): split MetagameAnalysisFrame into mixin package
Convert widgets/frames/metagame_analysis.py (619 LOC) into a package
following the match_history/timer_alert layout:

- __init__.py: public API (MetagameAnalysisFrame, main); the main
  entry-point is still reachable as ``widgets.frames.metagame_analysis:main``
  for the pyproject.toml console script
- frame.py: sys.path bootstrap, __init__, _build_ui, styling helpers
  (_stylize_button, _create_value_badge), and the standalone ``main()``
- handlers.py: MetagameAnalysisHandlersMixin -- on_format_change, the
  day / offset navigation callbacks, refresh_data worker chain
  (with _handle_error / _populate_data), _set_default_day_offset,
  update_visualization, _draw_pie_chart, _update_changes_display,
  _set_changes_html, _sync_navigation_controls, _set_busy, on_close
- properties.py: MetagameAnalysisPropertiesMixin -- translation (_t),
  aggregation (_aggregate_for_days), pure math (_calculate_percentages),
  color helper (_rgb_to_hex), HTML builders (_build_change_card,
  _build_changes_html)

Composition becomes:
  class MetagameAnalysisFrame(HandlersMixin, PropertiesMixin, wx.Frame): ...

Existing imports in widgets/handlers/app_event_handlers.py,
widgets/frames/app_frame.py, and the lazy-export map in
widgets/frames/__init__.py continue to resolve MetagameAnalysisFrame
via the package __init__, so no call sites change.

pyproject.toml's E402 per-file-ignore now points at the new
widgets/frames/metagame_analysis/frame.py where the sys.path bootstrap
lives.

Refs #427 (second pass, partial).
