+++ widgets/frames/metagame_analysis/__init__.py
@@ -1,0 +1,7 @@
+"""Metagame analysis viewer package."""
+
+from __future__ import annotations
+
+from widgets.frames.metagame_analysis.frame import MetagameAnalysisFrame, main
+
+__all__ = ["MetagameAnalysisFrame", "main"]
